Rename phosphor files and directories to lumino.
